backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
spacy==3.7.2
faiss-cpu==1.7.4
numpy==1.24.3
python-multipart==0.0.6
aiofiles==23.2.1
python-dateutil==2.8.2
